remove `"core/print/providers"` folder
- remove "core/print/providers::PrinterProvider" class
- remove "core/print/providers::PrinterProviderFactory" class
- remove "core/print/providers::PrinterQGISProvider" class
- refactor "core/print::PrintService" class (make use of "PrinterQGISProvider" as sole print service)
