--- src/app/core/print/providers/printerproviderfactory.js
@@ -1,20 +1,0 @@
-const {inherit, base} = require('core/utils/utils');
-const G3WObject = require('core/g3wobject');
-
-const PrinterQGISProvider = require('./qgis/printerQGISProvider');
-
-const Providers = {
-  'QGIS': PrinterQGISProvider
-};
-
-function PrinterProviderFactory() {
-  this.get = function(type = 'QGIS') {
-    return new Providers[type];
-  };
-  base(this);
-}
-
-inherit(PrinterProviderFactory, G3WObject);
-
-module.exports =  new PrinterProviderFactory;
-
